Move encryption and api functions into the base class
